backend/pyproject.toml
[project]
name = "golden-aura-backend"
version = "0.1.0"
description = "Django backend for Golden Aura Showroom"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.15,<4.0",
  "psycopg2-binary>=2.9,<3.0",
  "python-dotenv>=1.0,<2.0",
]

[tool.uv]
package = false
